Fix delete range command for integer overflow bug

+++ src/main/java/seedu/address/logic/parser/BulkIndexParserUtil.java
@@ -151,12 +151,12 @@ public final class BulkIndexParserUtil {
             throw new ParseException(MESSAGE_INVALID_RANGE);
         }
 
-        if (end - start + 1 > MAX_RANGE_SIZE) {
+        if ((long) end - (long) start + 1 > MAX_RANGE_SIZE) {
             throw new ParseException(MESSAGE_RANGE_TOO_LARGE);
         }
 
-        for (int i = start; i <= end; i++) {
-            indexSet.add(i);
+        for (long i = start; i <= end; i++) {
+            indexSet.add((int) i);
         }
     }
